+++ docs/source/history.rst
@@ -14,7 +14,7 @@ v1.1 series
 
     | **New features:**
 
-    * Added models: ``ApogeeKplerEchelle`` and ``ApokascEncoderDecoder``
+    * Added models: ``ApogeeKeplerEchelle`` and ``ApokascEncoderDecoder``
     * Input data can now be a dict, such as ``nn.train({'input': input_data, 'input': aux_input_data}, {'output': labels, 'output_aux': aux_labels})``
     * Added numerical integrator for NeuralODE
     * tqdm progress bar for model prediction